There were added category, exception handler, also improved UI, and capability to create an exel file of products

+++ src/main/java/dev/petproject/repository/ProductRepository.java
@@ -9,8 +9,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     @Query(value = "SELECT * FROM products p WHERE p.name LIKE %:keyword%", nativeQuery = true)
     List<Product> findProductByKeyword(@Param(value = "keyword") String keyword);
+
 }
